requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
reportlab>=4.0.0
click>=8.1.0
scikit-learn>=1.3.0
numba>=0.58.0
Cython>=3.0.0